feat(pagamento): implementa fluxo de pagamento com PIX, cartão e boleto, integra front com backend

+++ malune/src/main/java/com/example/malune/repository/ItemPedidoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
 
     List<ItemPedido> findByPedido(Pedido pedido);
 }
